+++ core/src/com/unciv/logic/automation/ConstructionAutomation.kt
@@ -2,7 +2,6 @@ package com.unciv.logic.automation
 
 import com.unciv.UncivGame
 import com.unciv.logic.city.CityConstructions
-import com.unciv.logic.city.INonPerpetualConstruction
 import com.unciv.logic.city.PerpetualConstruction
 import com.unciv.logic.civilization.CityAction
 import com.unciv.logic.civilization.NotificationIcon
@@ -194,7 +193,7 @@ class ConstructionAutomation(val cityConstructions: CityConstructions){
     }
 
     private fun addSpaceshipPartChoice() {
-        val spaceshipPart = buildableNotWonders.firstOrNull { it.uniques.contains("Spaceship part") }
+        val spaceshipPart = buildableNotWonders.firstOrNull { it.hasUnique(UniqueType.SpaceshipPart) }
         if (spaceshipPart != null) {
             var modifier = 1.5f
             if (preferredVictoryType == VictoryType.Scientific) modifier = 2f